fix(serving): coerce predict-input DataFrame to float dtype
Null-valued fields produced pandas object dtype columns, which
LightGBM rejects at score-time with 'pandas dtypes must be int, float
or bool'. Coercing the single-row DataFrame to float resolves it
(None -> np.nan in float dtype, LGBM's native-NaN path accepts that).
Same fix applied to the predict() utility used by the CLI.

+++ src/cardio_risk_rf/inference/predict.py
@@ -23,7 +23,7 @@ def predict(model: Any, features: dict[str, Any]) -> dict[str, Any]:
     """Run inference on a single feature mapping and return pred + class probabilities."""
     import pandas as pd
 
-    x = pd.DataFrame([features])
+    x = pd.DataFrame([features]).astype(float)
     proba = model.predict_proba(x)[0].tolist()
     pred = int(model.predict(x)[0])
     return {"pred": pred, "proba": proba}
